refactor: StoneEntity 패키지 변경
